# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Testing Shop API",
    "version": "1.0.0",
    "description": "Demo e-shop backend with fault injection for API/UI/Unit testing scenarios."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://{apiHost}",
      "description": "Deployment host",
      "variables": {
        "apiHost": {
          "default": "example.azurecontainerapps.io"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  },
...
  "paths": {
...
    "/faults/inject-error": {
      "get": {
        "tags": [
          "Faults"
        ],
        "summary": "Return intentional localized 400 response for network-fault training.",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "cs"
              ]
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Intentional bug response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}